refactor: presentation 계층으로 controller와 converter, dto 이동
