+++ src/test/java/com/techfork/recommendation/presentation/RecommendationConverterTest.java
@@ -1,7 +1,7 @@
-package com.techfork.domain.recommendation.converter;
+package com.techfork.recommendation.presentation;
 
 import com.techfork.post.domain.Post;
-import com.techfork.domain.recommendation.dto.RecommendedPostDto;
+import com.techfork.recommendation.presentation.response.RecommendedPostDto;
 import com.techfork.recommendation.domain.RecommendedPost;
 import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.useraccount.domain.User;
